\name{PDSCE-package}
\alias{PDSCE-package}
\alias{PDSCE}
\docType{package}
\title{
Positive definite sparse covariance estimators
}
\description{
A package to compute and tune some positive definite and sparse covariance estimators
}
\details{
\tabular{ll}{
Package: \tab PDSCE\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-08-23\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}

The main functions are \code{\link{pdsoft}}, \code{\link{pdsoft.cv}}, 
\code{\link{band.chol}}, and \code{\link{band.chol.cv}}. 
}
\author{
Adam J. Rothman

Maintainer: Adam J. Rothman <arothman@umn.edu>
}
\references{
Rothman, A. J., Levina, E., and Zhu, J. (2010).  A new approach to Cholesky-based covariance 
regularization in high dimensions. Biometrika 97(3): 539-550.

Rothman, A. J. (2012).  Positive definite estimators of large covariance matrices. Biometrika 99(3): 733-740
}


